libwsbm: resolve find_vma problem in kernel driver
BZ: 158272

When user space is 32 bit, kernel space is 64 bit,
user address space can above 3G.
While if we set address to uint64, need first convert address to unsigned long.x

+++ src/wsbm_ttmpool.c
@@ -548,7 +548,7 @@ ttm_pool_ub_create(struct _WsbmBufferPool *pool, unsigned long size, uint32_t pl
     arg.req.size = size;
     arg.req.placement = placement;
     arg.req.page_alignment = alignment / pageSize;
-    arg.req.user_address = user_ptr;
+    arg.req.user_address = (unsigned long)user_ptr;
 
     DRMRESTARTCOMMANDWRITEREAD(pool->fd, ttmPool->devOffset + TTM_PL_CREATE_UB,
 			       arg, ret);
